Remove circular dependency between material & mapview packages.
