+++ @here/harp-materials/lib/MapMeshMaterials.ts
@@ -4,13 +4,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { ViewRanges } from "@here/harp-datasource-protocol/lib/ViewRanges";
 import { applyMixinsWithoutProperties, chainCallbacks } from "@here/harp-utils";
-import { AnimatedExtrusionTileHandler } from "../../harp-mapview/lib/AnimatedExtrusionHandler";
 import { insertShaderInclude } from "./Utils";
 
 import * as THREE from "three";
 
-import { ViewRanges } from "../../harp-mapview/lib/ClipPlanesEvaluator";
 import extrusionShaderChunk from "./ShaderChunks/ExtrusionChunks";
 import fadingShaderChunk from "./ShaderChunks/FadingChunks";
 
@@ -559,6 +558,15 @@ export class FadingFeatureMixin implements FadingFeature {
 
 export namespace ExtrusionFeature {
     /**
+     * Minimum ratio value for extrusion effect
+     */
+    export const DEFAULT_RATIO_MIN: number = 0.001;
+    /**
+     * Maximum ratio value for extrusion effect
+     */
+    export const DEFAULT_RATIO_MAX: number = 1;
+
+    /**
      * Patch the THREE.ShaderChunk on first call with some extra shader chunks.
      */
     export function patchGlobalShaderChunks() {
@@ -581,7 +589,7 @@ export namespace ExtrusionFeature {
 
         if (
             extrusionMaterial.extrusionRatio !== undefined &&
-            extrusionMaterial.extrusionRatio >= AnimatedExtrusionTileHandler.DEFAULT_RATIO_MIN
+            extrusionMaterial.extrusionRatio >= ExtrusionFeature.DEFAULT_RATIO_MIN
         ) {
             // Add this define to differentiate it internally from other MeshBasicMaterial
             extrusionMaterial.defines.EXTRUSION_MATERIAL = "";
@@ -668,7 +676,7 @@ export namespace ExtrusionFeature {
             properties.shader.uniforms.extrusionRatio !== undefined
         ) {
             properties.shader.uniforms.extrusionRatio.value =
-                extrusionMaterial.extrusionRatio || AnimatedExtrusionTileHandler.DEFAULT_RATIO_MAX;
+                extrusionMaterial.extrusionRatio || ExtrusionFeature.DEFAULT_RATIO_MAX;
             extrusionMaterial.uniformsNeedUpdate = true;
         }
     }
@@ -684,7 +692,7 @@ export namespace ExtrusionFeature {
 export class ExtrusionFeatureMixin implements ExtrusionFeature {
     needsUpdate?: boolean;
     uniformsNeedUpdate?: boolean;
-    private m_extrusion: number = AnimatedExtrusionTileHandler.DEFAULT_RATIO_MAX;
+    private m_extrusion: number = ExtrusionFeature.DEFAULT_RATIO_MAX;
 
     /**
      * @see [[ExtrusionFeature#extrusion]]
@@ -821,7 +829,7 @@ export class MapMeshBasicMaterial extends THREE.MeshBasicMaterial
     }
 
     get extrusionRatio(): number {
-        return AnimatedExtrusionTileHandler.DEFAULT_RATIO_MAX;
+        return ExtrusionFeature.DEFAULT_RATIO_MAX;
     }
     // tslint:disable-next-line:no-unused-variable
     set extrusionRatio(value: number) {
@@ -953,7 +961,7 @@ export class MapMeshStandardMaterial extends THREE.MeshStandardMaterial
     }
 
     get extrusionRatio(): number {
-        return AnimatedExtrusionTileHandler.DEFAULT_RATIO_MAX;
+        return ExtrusionFeature.DEFAULT_RATIO_MAX;
     }
     // tslint:disable-next-line:no-unused-variable
     set extrusionRatio(value: number) {
